#172 add GeneralInfoService
